+++ boardgamegeek/api.py
@@ -1,5 +1,4 @@
 # coding: utf-8
-
 """
 :mod:`boardgamegeek.api` - Core functions
 =========================================
@@ -9,7 +8,7 @@ objects.
 
 .. module:: boardgamegeek.api
    :platform: Unix, Windows
-   :synopsis: module handling communication with the online BGG API
+   :synopsis: module handling communication with the online BoardGameGeek API
 
 .. moduleauthor:: Cosmin Luță <q4break@gmail.com>
 """
@@ -49,23 +48,21 @@ HOT_ITEM_CHOICES = ["boardgame", "rpg", "videogame", "boardgameperson", "rpgpers
 
 
 class BoardGameGeekNetworkAPI(object):
+    """
+    Base class for the BoardGameGeek websites APIs. All site-specific clients are derived from this.
 
+    :param str api_endpoint: URL of the API
+    :param str cache: URL indicating the cache to use, or ``None`` if caching should be disabled (not recommended)
+    :param integer timeout: timeout for a request
+    :param integer retries: how many retries to perform in special cases
+    :param integer retry_delay: delay between retries (seconds)
+    """
     SEARCH_RPG_ITEM = 1
     SEARCH_VIDEO_GAME = 2
     SEARCH_BOARD_GAME = 4
     SEARCH_BOARD_GAME_EXPANSION = 8
 
     def __init__(self, api_endpoint, cache, timeout, retries, retry_delay, requests_per_minute):
-        """
-
-        :param api_endpoint: URL of the API
-        :param cache: if non-empty, use the cache described by this URI
-        :param timeout: timeout for a request
-        :param retries: how many retries to perform in special cases
-        :param retry_delay: delay between retries (seconds)
-        :return:
-        """
-
         self._search_api_url = api_endpoint + "/search"
         self._thing_api_url = api_endpoint + "/thing"
         self._guild_api_url = api_endpoint + "/guild"
@@ -85,62 +82,53 @@ class BoardGameGeekNetworkAPI(object):
         # add the rate limiting adapter
         self.requests_session.mount(api_endpoint, RateLimitingAdapter(rpm=requests_per_minute))
 
-    def _get_game_id(self, name, game_type, first=False):
+    def _get_game_id(self, name, game_type, choose):
         """
         Returns the BGG ID of a game, searching by name
 
-        :param name: The name of the game to search for
-        :param game_type: the game type: "rpgitem", "videogame", "boardgame", "boardgameexpansion"
-        :param first: if true, return the first result, otherwise return the most recent (by year published)
+        :param str name: the name of the game to search for
+        :param str game_type: the game type ("rpgitem", "videogame", "boardgame", "boardgameexpansion")
+        :param str choose: method of selecting the game by name, when dealing with multiple results. Valid values are "first", "recent" or "best-rank"
         :return: ``None`` if game wasn't found
-        :return: integer value of the game's id
+        :return: game's id
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekError` in case of invalid name
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
         """
 
-        if game_type not in ["rpgitem", "videogame", "boardgame", "boardgameexpansion"]:
-            raise BoardGameGeekError("invalid game type: {}".format(game_type))
+        if choose not in ["first", "recent", "best-rank"]:
+            raise BoardGameGeekError("invalid value for parameter 'choose': {}".format(choose))
 
         log.debug("getting game id for '{}'".format(name))
+        res = self.search(name, search_type=[game_type], exact=True)
 
-        try:
-            root = get_parsed_xml_response(self.requests_session,
-                                           self._search_api_url,
-                                           params={"query": name, "exact": 1},
-                                           timeout=self._timeout,
-                                           retries=self._retries,
-                                           retry_delay=self._retry_delay)
-        except BoardGameGeekAPINonXMLError:
+        if not res:
             return None
 
-        # in case there are multiple results, try to find the most recent one
-        _year = None
-        game_id = None
-
-        # game_type can be rpgitem, videogame, boardgame, or boardgameexpansion
-        for game in root.findall(".//item[@type='{}']".format(game_type)):
-            year = xml_subelement_attr(game, "yearpublished", convert=int, default=0, quiet=True)
-            if _year is None:
-                _year = year
-                game_id = int(game.attrib.get("id"))
-                if first:
-                    # if we want the first result, break
-                    break
-            elif year > _year:
-                _year = year
-                game_id = int(game.attrib.get("id"))
-
-        if _year is None:
-            log.warn("game not found: {}".format(name))
-            return None
-
-        return game_id
+        if choose == "first":
+            return res[0].id
+        elif choose == "recent":
+            return max(res, key=lambda x: x.year if x.year is not None else -300000).id
+        else:
+            # getting the best rank requires fetching the data of all games returned
+            game_data = [self.game(game_id=r.id) for r in res]
+            # ...and selecting the one with the best ranking
+            return min(game_data, key=lambda x: x.boardgame_rank if x.boardgame_rank is not None else 10000000000).id
 
     def guild(self, guild_id, progress=None):
         """
         Retrieves details about a guild
 
-        :param guild_id: The ID of the guild
-        :param progress: Optional progress callback for member fetching
-        :return: :class:`boardgamegeek.guild.Guild` object containing the details
+        :param integer guild_id: the id number of the guild
+        :param callable progress: an optional callable for reporting progress, taking two integers (``current``, ``total``) as arguments
+        :return: ``Guild`` object containing the data
+        :return: ``None`` if the information couldn't be retrieved
+        :rtype: :py:class:`boardgamegeek.guild.Guild`
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekError` in case of an invalid guild id
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
         """
 
         try:
@@ -225,12 +213,17 @@ class BoardGameGeekNetworkAPI(object):
         """
         Retrieves details about an user
 
-        :param name: user's login name
-        :return: :py:class:`boardgamegeek.user.User`
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekError` in case of invalid user name
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
+        :param str name: user's login name
+        :param callable progress: an optional callable for reporting progress when fetching the buddy list/guilds, taking two integers (``current``, ``total``) as arguments
+
+        :return: ``User`` object
+        :rtype: :py:class:`boardgamegeek.user.User`
+        :return: ``None`` if the user couldn't be found
+
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekError` in case of invalid user name
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
         """
 
         if not name:
@@ -345,19 +338,21 @@ class BoardGameGeekNetworkAPI(object):
 
     def plays(self, name=None, game_id=None, progress=None, min_date=None, max_date=None):
         """
-        Retrieves the user's play list
+        Retrieves the plays for an user (if using ``name``) or for a game (if using ``game_id``)
 
-        :param name: user name to retrieve the plays for
-        :param game_id: game id to retrieve the plays for
-        :param min_date: return only plays of the specified date or later.
-        :param max_date: return only plays of the specified date or earlier.
-        :type mindate: datetime.date
-        :type max_date: datetime.date
-        :return: :py:class:`boardgamegeek.plays.Plays` object containing all the plays
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekError` on errors
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
+        :param str name: user name to retrieve the plays for
+        :param integer game_id: game id to retrieve the plays for
+        :param callable progress: an optional callable for reporting progress, taking two integers (``current``, ``total``) as arguments
+        :param datetime.date min_date: return only plays of the specified date or later.
+        :param datetime.date max_date: return only plays of the specified date or earlier.
+
+        :return: object containing all the plays
+        :rtype: :py:class:`boardgamegeek.plays.Plays`
+        :return: ``None`` if the user/game couldn't be found
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekError` on errors
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
 
         """
         if not name and not game_id:
@@ -483,12 +478,16 @@ class BoardGameGeekNetworkAPI(object):
         """
         Return the list of "Hot Items"
 
-        :param item_type: hot item type (valid values: "boardgame", "rpg", "videogame", "boardgameperson", "rpgperson", "boardgamecompany", "rpgcompany", "videogamecompany")
-        :return: :py:class:`boardgamegeek.hotitems.HotItems` containing the hot items
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekError` if the parameter is invalid
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
+        :param str item_type: hot item type. Valid values: "boardgame", "rpg", "videogame", "boardgameperson", "rpgperson", "boardgamecompany", "rpgcompany", "videogamecompany")
+
+        :return: ``HotItems`` object
+        :rtype: :py:class:`boardgamegeek.hotitems.HotItems`
+        :return: ``None`` in case the hot items couldn't be retrieved
+
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekError` if the parameter is invalid
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
         """
         if item_type not in HOT_ITEM_CHOICES:
             raise BoardGameGeekError("invalid type specified")
@@ -522,14 +521,15 @@ class BoardGameGeekNetworkAPI(object):
         """
         Returns the user's game collection
 
-        :param user_name: user name to retrieve the collection for
-        :return: :py:class:`boardgamegeek.collection.Collection`
+        :param str user_name: user name to retrieve the collection for
+        :return: ``Collection`` object
+        :rtype: :py:class:`boardgamegeek.collection.Collection`
         :return: ``None`` if user not found
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIError` if there was a problem getting the collection
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekError` in case of invalid parameters
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
+
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekError` in case of invalid parameters
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
         """
         if not user_name:
             raise BoardGameGeekError("no user name specified")
@@ -584,14 +584,16 @@ class BoardGameGeekNetworkAPI(object):
         """
         Search for a game
 
-        :param query: the string to search for
-        :param search_type: list of strings indicating what to search for. Valid contained values are: "rpgitem", "videogame", "boardgame", "boardgameexpansion"
-        :param exact: if True, try to match the name exactly
-        :return: list of :py:class:`boardgamegeek.search.SearchResult` objects
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekError` in case of invalid query
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
+        :param str query: the string to search for
+        :param str search_type: list of strings indicating what to search for. Valid contained values are: "rpgitem", "videogame", "boardgame", "boardgameexpansion"
+        :param bool exact: if True, try to match the name exactly
+        :return: list of ``SearchResult``
+        :rtype: list of :py:class:`boardgamegeek.search.SearchResult`
+
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekError` in case of invalid query
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
         """
         if not query:
             raise BoardGameGeekError("invalid query string")
@@ -642,7 +644,7 @@ class BoardGameGeekNetworkAPI(object):
         for item in root.findall("item"):
             kwargs = {"id": item.attrib["id"],
                       "name": xml_subelement_attr(item, "name"),
-                      "yearpublished": xml_subelement_attr(item, "yearpublished", convert=int, quiet=True),
+                      "yearpublished": xml_subelement_attr(item, "yearpublished", default=0, convert=int, quiet=True),
                       "type": item.attrib["type"]}
 
             results.append(SearchResult(kwargs))
@@ -652,10 +654,17 @@ class BoardGameGeekNetworkAPI(object):
 
 class BoardGameGeek(BoardGameGeekNetworkAPI):
     """
-        Python interface for www.boardgamegeek.com's XML API.
+        Python interface for www.boardgamegeek.com's XML API 2.
 
         Caching for the requests can be used by specifying an URI for the ``cache`` parameter. By default, an in-memory
-        cache is used, with "sqlite" being the other currently supported option.
+        cache is used, with sqlite being the other currently supported option.
+
+        :param cache: URL indicating the cache to use for HTTP requests, ``None`` if disabled
+        :param timeout: Timeout for network operations
+        :param retries: Number of retries to perform in case the API returns HTTP 202 (retry) or in case of timeouts
+        :param retry_delay: Time to sleep between retries when the API returns HTTP 202 (retry)
+        :param disable_ssl: If true, use HTTP instead of HTTPS for calling the BGG API
+        :param requests_per_minute: how many requests per minute to allow to go out to BGG (throttle prevention)
 
         Example usage::
 
@@ -668,15 +677,7 @@ class BoardGameGeek(BoardGameGeekNetworkAPI):
 
     """
     def __init__(self, cache="memory:///?ttl=3600", timeout=15, retries=3, retry_delay=5, disable_ssl=False, requests_per_minute=DEFAULT_REQUESTS_PER_MINUTE):
-        """
 
-        :param cache: Cache to use for requests, ``None`` if disabled
-        :param timeout: Timeout for network operations
-        :param retries: Number of retries to perform in case the API returns HTTP 202 (retry) or in case of timeouts
-        :param retry_delay: Time to sleep between retries when the API returns HTTP 202 (retry)
-        :param disable_ssl: If true, use HTTP instead of HTTPS for calling the BGG API
-        :param requests_per_minute: how many requests per minute to allow to go out to BGG (throttle prevention)
-        """
         api_endpoint = "http{}://www.boardgamegeek.com/xmlapi2".format("" if disable_ssl else "s")
         super(BoardGameGeek, self).__init__(api_endpoint=api_endpoint,
                                             cache=cache,
@@ -685,41 +686,44 @@ class BoardGameGeek(BoardGameGeekNetworkAPI):
                                             retry_delay=retry_delay,
                                             requests_per_minute=requests_per_minute)
 
-    def get_game_id(self, name):
+    def get_game_id(self, name, choose="first"):
         """
         Returns the BGG ID of a game, searching by name
 
-        :param name: The name of the game to search for
-        :param game_type: the game type: "rpgitem", "videogame", "boardgame", "boardgameexpansion"
-        :param first: if true, return the first result, otherwise return the most recent (by year published)
+        :param str name: The name of the game to search for
+        :param str choose: method of selecting the game by name, when dealing with multiple results. Valid values are "first", "recent" or "best-rank"
+        :return: the game's id
+        :rtype: integer
         :return: ``None`` if game wasn't found
-        :return: integer value of the game's id
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekError` in case of invalid name
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
         """
-        return self._get_game_id(name, "boardgame")
+        return self._get_game_id(name, game_type="boardgame", choose=choose)
 
-    def game(self, name=None, game_id=None):
+    def game(self, name=None, game_id=None, choose="first"):
         """
         Get information about a game.
 
-        WARNING: it can happen that there are multiple games with the same name, and this function might not always
-        pick up the one you're interested in. As a workaround, call the `boardgamegeek.api.BoardGame.search` function
-        to get the exact id, then use it here.
-
-        :param name: If not None, get information about a game with this name
-        :param game_id:  If not None, get information about a game with this id
-        :return: :py:class:`boardgamegeek.games.BoardGame`
+        :param str name: If not None, get information about a game with this name
+        :param integer game_id:  If not None, get information about a game with this id
+        :param str choose: method of selecting the game by name, when dealing with multiple results. Valid values are "first", "recent" or "best-rank"
+        :return: ``BoardGame`` object
+        :rtype: :py:class:`boardgamegeek.games.BoardGame`
         :return: ``None`` if the game wasn't found
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekError` in case of invalid name and id
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
+
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekError` in case of invalid name or game_id
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
         """
 
         if not name and game_id is None:
             raise BoardGameGeekError("game name or id not specified")
 
         if game_id is None:
-            game_id = self.get_game_id(name)
+            game_id = self.get_game_id(name, choose=choose)
             if game_id is None:
                 log.error("couldn't find any game named '{}'".format(name))
                 return None
@@ -820,13 +824,13 @@ class BoardGameGeek(BoardGameGeekNetworkAPI):
         """
         Return a list containing all games with the given name
 
-        :param name: The name of the game to search for
+        :param str name: the name of the game to search for
         :return: list of :py:class:`boardgamegeek.games.BoardGame`
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
-        :raises: :py:class:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIRetryError` if this request should be retried after a short delay
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekAPIError` if the response couldn't be parsed
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekTimeoutError` if there was a timeout
         """
         return [self.game(game_id=s.id)
                 for s in self.search(name,
-                                     search_type=BoardGameGeekNetworkAPI.SEARCH_BOARD_GAME | BoardGameGeekNetworkAPI.SEARCH_BOARD_GAME_EXPANSION,
+                                     search_type=["boardgame", "boardgameexpansion"],
                                      exact=True)]