+++ boardgamegeek/exceptions.py
@@ -1,3 +1,15 @@
+# coding: utf-8
+"""
+:mod:`boardgamegeek.exceptions` - Exceptions
+============================================
+
+.. module:: boardgamegeek.exceptions
+   :platform: Unix, Windows
+   :synopsis: exceptions used in the package
+
+.. moduleauthor:: Cosmin Luță <q4break@gmail.com>
+"""
+
 # boardgamegeek library base exception
 class BoardGameGeekError(Exception):
     pass