+++ boardgamegeek/games.py
@@ -1,3 +1,15 @@
+# coding: utf-8
+"""
+:mod:`boardgamegeek.games` - Games information
+==============================================
+
+.. module:: boardgamegeek.games
+   :platform: Unix, Windows
+   :synopsis: classes for storing games information
+
+.. moduleauthor:: Cosmin Luță <q4break@gmail.com>
+
+"""
 from __future__ import unicode_literals
 from copy import copy
 
@@ -8,9 +20,8 @@ from .utils import fix_url
 
 class CollectionBoardGame(Thing):
     """
-    A boardgame retrieved from the collection information, which has
-    less information than the one retrieved via the /thing api and which
-    also contains some user-specific information
+    A boardgame retrieved from the collection information, which has less information than the one retrieved
+    via the /thing api and which also contains some user-specific information.
     """
     def __repr__(self):
         return "CollectionBoardGame (id: {})".format(self.id)
@@ -37,49 +48,95 @@ class CollectionBoardGame(Thing):
         return self._data.get("lastmodified")
 
     @property
+    def last_modified(self):
+        """
+        :return: last modified date
+        :rtype: str
+        """
+        return self._data.get("lastmodified")
+
+    @property
     def rating(self):
+        """
+        :return: game rating
+        :rtype: float
+        :return: ``None`` if n/a
+        """
         return self._data.get("rating")
 
     @property
     def owned(self):
+        """
+        :return: game owned
+        :rtype: bool
+        """
         return bool(int(self._data.get("own", 0)))
 
     @property
     def preordered(self):
+        """
+        :return: game preordered
+        :rtype: bool
+        """
         return bool(int(self._data.get("preordered", 0)))
 
     @property
     def prev_owned(self):
+        """
+        :return: game previously owned
+        :rtype: bool
+        """
         return bool(int(self._data.get("prevowned", 0)))
 
     @property
     def want(self):
+        """
+        :return: game wanted
+        :rtype: bool
+        """
         return bool(int(self._data.get("want", 0)))
 
     @property
     def want_to_buy(self):
+        """
+        :return: want to buy
+        :rtype: bool
+        """
         return bool(int(self._data.get("wanttobuy", 0)))
 
     @property
     def want_to_play(self):
+        """
+        :return: want to play
+        :rtype: bool
+        """
         return bool(int(self._data.get("wanttoplay", 0)))
 
     @property
     def for_trade(self):
+        """
+        :return: game for trading
+        :rtype: bool
+        """
         return bool(int(self._data.get("fortrade", 0)))
 
     @property
     def wishlist(self):
+        """
+        :return: game on wishlist
+        :rtype: bool
+        """
         return bool(int(self._data.get("wishlist", 0)))
 
     @property
     def wishlist_priority(self):
+        # TODO: convert to int (it's str)
         return self._data.get("wishlistpriority")
 
 
 class BoardGame(Thing):
     """
-    An object containing the core information about a game.
+    Object containing information about a boardgame
     """
     def __init__(self, data):
 
@@ -136,6 +193,12 @@ class BoardGame(Thing):
         return "BoardGame (id: {})".format(self.id)
 
     def add_expanded_game(self, data):
+        """
+        Add a game expanded by this one
+
+        :param dict data: expanded game's data
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekError` if data is invalid
+        """
         try:
             if data["id"] not in self._expands_set:
                 self._data["expands"].append(data)
@@ -145,6 +208,12 @@ class BoardGame(Thing):
             raise BoardGameGeekError("invalid expanded game data")
 
     def add_expansion(self, data):
+        """
+        Add an expansion of this game
+
+        :param dict data: expansion data
+        :raises: :py:exc:`boardgamegeek.exceptions.BoardGameGeekError` if data is invalid
+        """
         try:
             if data["id"] not in self._expansions_set:
                 self._data["expansions"].append(data)
@@ -228,139 +297,277 @@ class BoardGame(Thing):
 
     @property
     def alternative_names(self):
-        return self._data.get("alternative_names")
+        """
+        :return: alternative names
+        :rtype: list of str
+        """
+        return self._data.get("alternative_names", [])
 
     @property
     def thumbnail(self):
+        """
+        :return: thumbnail URL
+        :rtype: str
+        :return: ``None`` if n/a
+        """
         return self._data.get("thumbnail")
 
     @property
     def image(self):
+        """
+        :return: image URL
+        :rtype: str
+        :return: ``None`` if n/a
+        """
         return self._data.get("image")
 
     @property
     def description(self):
-        return self._data.get("description")
+        """
+        :return: description
+        :rtype: str
+        """
+        return self._data.get("description", "")
 
     @property
     def families(self):
-        return self._data.get("families")
+        """
+        :return: families
+        :rtype: list of str
+        """
+        return self._data.get("families", [])
 
     @property
     def categories(self):
-        return self._data.get("categories")
+        """
+        :return: categories
+        :rtype: list of str
+        """
+        return self._data.get("categories", [])
 
     @property
     def mechanics(self):
-        return self._data.get("mechanics")
+        """
+        :return: mechanics
+        :rtype: list of str
+        """
+        return self._data.get("mechanics", [])
 
     @property
     def expansions(self):
         """
-
-        :return: list of expansions for this item
+        :return: expansions
+        :rtype: list of :py:class:`boardgamegeek.things.Thing`
         """
         return self._expansions
 
     @property
     def expands(self):
         """
-        :return: list of games this item expands
+        :return: games this item expands
+        :rtype: list of :py:class:`boardgamegeek.things.Thing`
         """
         return self._expands
 
     @property
     def implementations(self):
-        return self._data.get("implementations")
+        """
+        :return: implementations
+        :rtype: list of str
+        """
+        return self._data.get("implementations", [])
 
     @property
     def designers(self):
-        return self._data.get("designers")
+        """
+        :return: designers
+        :rtype: list of str
+        """
+        return self._data.get("designers", [])
 
     @property
     def artists(self):
-        return self._data.get("artists")
+        """
+        :return: artists
+        :rtype: list of str
+        """
+        return self._data.get("artists", [])
 
     @property
     def publishers(self):
-        return self._data.get("publishers")
+        """
+        :return: publishers
+        :rtype: list of str
+        """
+        return self._data.get("publishers", [])
 
     @property
     def expansion(self):
         """
-
-        :return: True if this item is an expansion to some other item
+        :return: True if this item is an expansion
+        :rtype: bool
         """
         return self._data.get("expansion", False)
 
     @property
     def year(self):
+        """
+        :return: publishing year
+        :rtype: integer
+        :return: ``None`` if n/a
+        """
         return self._data.get("yearpublished")
 
     @property
     def min_players(self):
+        """
+        :return: minimum number of players
+        :rtype: integer
+        :return: ``None`` if n/a
+        """
         return self._data.get("minplayers")
 
     @property
     def max_players(self):
+        """
+        :return: maximum number of players
+        :rtype: integer
+        :return: ``None`` if n/a
+        """
         return self._data.get("maxplayers")
 
     @property
     def playing_time(self):
+        """
+        :return: playing time
+        :rtype: integer
+        :return: ``None`` if n/a
+        """
         return self._data.get("playingtime")
 
     @property
     def min_age(self):
+        """
+        :return: minimum recommended age
+        :rtype: integer
+        :return: ``None`` if n/a
+        """
         return self._data.get("minage")
 
     @property
     def users_rated(self):
+        """
+        :return: how many users rated the game
+        :rtype: integer
+        :return: ``None`` if n/a
+        """
         return self._data.get("usersrated")
 
     @property
     def rating_average(self):
+        """
+        :return: average rating
+        :rtype: float
+        :return: ``None`` if n/a
+        """
         return self._data.get("average")
 
     @property
     def rating_bayes_average(self):
+        """
+        :return: bayes average rating
+        :rtype: float
+        :return: ``None`` if n/a
+        """
         return self._data.get("bayesaverage")
 
     @property
     def rating_stddev(self):
+        """
+        :return: standard deviation
+        :rtype: float
+        :return: ``None`` if n/a
+        """
         return self._data.get("stddev")
 
     @property
     def rating_median(self):
+        """
+        :return:
+        :rtype: float
+        :return: ``None`` if n/a
+        """
         return self._data.get("median")
 
     @property
     def users_owned(self):
+        """
+        :return: number of users owning this game
+        :rtype: integer
+        :return: ``None`` if n/a
+        """
         return self._data.get("owned")
 
     @property
     def users_trading(self):
+        """
+        :return: number of users trading this game
+        :rtype: integer
+        :return: ``None`` if n/a
+        """
         return self._data.get("trading")
 
     @property
     def users_wanting(self):
+        """
+        :return: number of users wanting this game
+        :rtype: integer
+        :return: ``None`` if n/a
+        """
         return self._data.get("wanting")
 
     @property
     def users_wishing(self):
+        """
+        :return: number of users wishing for this game
+        :rtype: integer
+        :return: ``None`` if n/a
+        """
         return self._data.get("wishing")
 
     @property
     def users_commented(self):
+        """
+        :return: number of user comments
+        :rtype: integer
+        :return: ``None`` if n/a
+        """
         return self._data.get("numcomments")
 
     @property
     def rating_num_weights(self):
+        """
+        :return:
+        :rtype: integer
+        :return: ``None`` if n/a
+        """
         return self._data.get("numweights")
 
     @property
     def rating_average_weight(self):
+        """
+        :return: average weight
+        :rtype: float
+        :return: ``None`` if n/a
+        """
         return self._data.get("averageweight")
 
     @property
     def ranks(self):
+        """
+        :return: rankings of this game
+        :rtype: list of dicts, keys: ``friendlyname`` (the friendly name of the rank, e.g. "Board Game Rank"), ``name``
+                (name of the rank, e.g "boardgame"), ``value`` (the rank)
+        :return: ``None`` if n/a
+        """
         return self._data.get("ranks")